+++ model/client.go
@@ -29,7 +29,7 @@ type Clients []*Client
 
 // Client represents a client model definition.
 type Client struct {
-	ID        int64     `json:"id" gorm:"primary_key"`
+	ID        uint      `json:"id" gorm:"primary_key"`
 	Slug      string    `json:"slug" sql:"unique_index"`
 	Name      string    `json:"name" sql:"unique_index"`
 	Value     string    `json:"uuid" sql:"unique_index"`
