+++ model/key.go
@@ -29,7 +29,7 @@ type Keys []*Key
 
 // Key represents a key model definition.
 type Key struct {
-	ID        int64     `json:"id" gorm:"primary_key"`
+	ID        uint      `json:"id" gorm:"primary_key"`
 	Slug      string    `json:"slug" sql:"unique_index"`
 	Name      string    `json:"name" sql:"unique_index"`
 	Value     string    `json:"key" sql:"unique_index"`
