+++ model/minecraft.go
@@ -19,7 +19,7 @@ type Minecrafts []*Minecraft
 
 // Minecraft represents a minecraft model definition.
 type Minecraft struct {
-	ID        int64     `json:"id" gorm:"primary_key"`
+	ID        uint      `json:"id" gorm:"primary_key"`
 	Slug      string    `json:"slug" sql:"unique_index"`
 	Name      string    `json:"version" sql:"unique_index"`
 	Type      string    `json:"type"`
