+++ model/user.go
@@ -35,7 +35,7 @@ type Users []*User
 
 // User represents a user model definition.
 type User struct {
-	ID         int64       `json:"id" gorm:"primary_key"`
+	ID         uint        `json:"id" gorm:"primary_key"`
 	Permission *Permission `json:"permission"`
 	Slug       string      `json:"slug" sql:"unique_index"`
 	Username   string      `json:"username" sql:"unique_index"`
